Add task for generating javascript parsers
Added a gradle task for generating the javascript parsers.  By default, they will be generated to `build/javascript/parsers`, but this can be overridden on the commandline with `-Pshr.js.parsers.dir=/path/to/dir`.

Code for using these parsers to generate JSON representations will be developed in a separate repository.