Lexer/Parser changes for mapping grammar
Better support for mapping to arbitrary targets -- with the target phrases being treated as long strings of characters.  This is accomplished by using lexer modes.

This simplifies the parser a bit too as we no longer distinguish between FieldToFieldMappings and FieldToURLMappings, nor do we try to parse targets into parts.  That's up the the mapping tool to decide how it wants to treat the target phases.