Remove "with code" type constructs from grammar
Instead of having to say "Foo with CodeableConcept from Bar", you can now say "Foo from Bar".  The expander will take care of the details regarding getting the constraint path right (and pointing to a code-ish thing).

For temporary backwards compatibility, I've left in an optional partial that will allow the "with code" type phrases, but the importer ignores them.