WIP: Constraining types to choice
This modifies the grammar to allow "Foo is type Bar or Baz".  It's unclear if this will be the final grammar construct or if it will be slightly different.

Also note that this *reintroduces* choices into fields (when you say a superclass field can be a choice of specific subtypes), so has some ramifications we need to consider.